Move this getResumeList
