+++ src/components/App/App.jsx
@@ -6,8 +6,7 @@ import Projects from "../projects/Projects";
 import React from "react";
 import ResumeDetails from "../resumes/Details";
 import ResumeEntry from "../resumes/Entry";
-import ResumeList from "../resumes/ResumeList/ResumeList";
-import { Typography } from "@material-ui/core";
+import ResumeList from "../resumes/ResumeList";
 import { Route, BrowserRouter as Router, Switch } from "react-router-dom";
 import "./App.scss";
